+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Sport form - update redirection
   - Leagues - Add « add league » menu link action
   - Day - remove alias generation (please use pathauto config)
+  - Extract static method to service
 
 ## 8.x-1.0-alpha15, XXXX-XX-XX
   - Ranking - add average ranking